Add files via upload
This is just a simple bash script to display the temperatures of all the major cities in India!

Steps to run the script !

1) To run the program on Windows simply download and extract the zip file. 
2) Open the extracted folder and click on the path of the folder 
3) Type and enter cmd. (You can also open the command prompt via other means and navigate to the extracted folder path)
4) You can run the script by typing "bash for_weather.sh" and enter
5) Or you can type "chmod +x for_weather.sh" and enter (This line of code is to give permission to execute the script)
6) Now type in "./for_weather.sh" and enter !

There you have it. You can find the temperatures of all the major cities in India through this script. 
Please Note: This is just a basic script that is completely beginner level. There's a long way to go!